Make logging tests less coupled to the implementation
By using mocks we created a tight coupling between the implementation
and the tests. For example, logging a message on a debug level could be
done in several ways:

  logger.debug("message")
  logger.add(Logger::DEBUG, "message")
  logger.log(Logger::DEBUG, "message")
  ...

The tests should not care which way the implementation uses and just
check the resulting log. This commit makes them do exactly that.